requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.1
idna==2.8
multidict==4.7.4
PyYAML==5.3
requests==2.22.0
tinydb==3.15.2
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
